factor out material logic, keep private variables
