+++ src/Engines/Extensions/index.ts
@@ -1,2 +1,3 @@
 export * from "./engine.occlusionQuery";
 export * from "./engine.transformFeedback";
+export * from "./engine.multiview";
